# Conflicts:
#	project/src/transpiler/transpile_helper.cpp
#	project/src/transpiler/vardecl.cpp

+++ project/src/transpiler/recorddecl.cpp
@@ -1,9 +1,13 @@
 #include "recorddecl.h"
-#include "unit_transpiler.h"
+
+#include <algorithm>
+
 #include "transpile_helper.h"
+#include "unit_transpiler.h"
 #include "vardecl.h"
 
-std::vector<RecordType> ProcessRecordType(const clang::RecordDecl *RD, bool is_local) {
+std::vector<RecordType> ProcessRecordType(const clang::RecordDecl *RD,
+                                          bool is_local) {
   extern UnitTranspiler transpiler;
   std::vector<RecordType> types;
   RecordType *existed = transpiler.record_manager_.GetById(RD->getID());
@@ -21,17 +25,19 @@ std::vector<RecordType> ProcessRecordType(const clang::RecordDecl *RD, bool is_l
   if (RD->hasNameForLinkage()) {
     name += RD->getNameAsString();
   } else {
-    name += std::to_string(reinterpret_cast<uint64_t>(RD)); // NOLINT(cppcoreguidelines-pro-type-reinterpret-cast)
+    name += std::to_string(reinterpret_cast<uint64_t>(
+        RD));  // NOLINT(cppcoreguidelines-pro-type-reinterpret-cast)
   }
 
-  size_t size = 0;
+  uint64_t size = 0;
 
   std::map<std::string, size_t> fields;
   size_t shift = 0;
 
   for (auto it = RD->decls_begin(); it != RD->decls_end(); it++) {
     if (it->getKind() == clang::Decl::Record) {
-      auto new_types = ProcessRecordType(llvm::dyn_cast<clang::RecordDecl>(*it), is_local);
+      auto new_types =
+          ProcessRecordType(llvm::dyn_cast<clang::RecordDecl>(*it), is_local);
       types.insert(types.end(), new_types.begin(), new_types.end());
     } else if (it->getKind() == clang::Decl::Field) {
       auto *field = llvm::dyn_cast<clang::FieldDecl>(*it);
@@ -54,6 +60,7 @@ std::vector<RecordType> ProcessRecordType(const clang::RecordDecl *RD, bool is_l
       }
     }
   }
-  types.push_back(transpiler.record_manager_.Add(RD->getID(), name, size, fields, is_local));
+  types.push_back(transpiler.record_manager_.Add(RD->getID(), name, size,
+                                                 fields, is_local));
   return types;
 }
